+++ Kernel/System/DynamicField/Driver/ProcessManagement/ActivityID.pm
@@ -20,6 +20,7 @@ our @ObjectDependencies = (
     'Kernel::System::DynamicFieldValue',
     'Kernel::System::Main',
     'Kernel::System::ProcessManagement::Activity',
+    'Kernel::System::ProcessManagement::DB::Activity',
     'Kernel::System::Ticket::ColumnFilter',
 );
 
@@ -177,6 +178,90 @@ sub ColumnFilterValuesGet {
     }
 
     return $ColumnFilterValues;
+}
+
+sub SearchFieldParameterBuild {
+    my ( $Self, %Param ) = @_;
+
+    # Get field value.
+    my $Value = $Self->SearchFieldValueGet(%Param);
+
+    # Set operator.
+    my $Operator = 'Equals';
+
+    # Search for a wild card in the value.
+    if ( $Value && ( $Value =~ m{\*} || $Value =~ m{\|\|} ) ) {
+
+        # Change operator.
+        $Operator = 'Like';
+    }
+
+    if ( $Param{DynamicFieldConfig}->{Name} eq 'ProcessManagementActivityID' && $Value ) {
+
+        my $ActivityEntityIDs = $Kernel::OM->Get('Kernel::System::ProcessManagement::DB::Activity')->ActivitySearch(
+            ActivityName => $Value,
+        );
+
+        if ( IsArrayRefWithData($ActivityEntityIDs) ) {
+
+            # Add search term from input field.
+            push @{$ActivityEntityIDs}, $Value;
+
+            # Return search parameter structure.
+            return {
+                Parameter => {
+                    $Operator => $ActivityEntityIDs,
+                },
+                Display => $Value,
+            };
+        }
+    }
+
+    # Return search parameter structure.
+    return {
+        Parameter => {
+            $Operator => $Value,
+        },
+        Display => $Value,
+    };
+}
+
+sub StatsSearchFieldParameterBuild {
+    my ( $Self, %Param ) = @_;
+
+    my $Value = $Param{Value};
+
+    # set operator
+    my $Operator = 'Equals';
+
+    # search for a wild card in the value
+    if ( $Value && $Value =~ m{\*} ) {
+
+        # change operator
+        $Operator = 'Like';
+    }
+
+    if ( $Param{DynamicFieldConfig}->{Name} eq 'ProcessManagementActivityID' && $Value ) {
+
+        my $ActivityEntityIDs = $Kernel::OM->Get('Kernel::System::ProcessManagement::DB::Activity')->ActivitySearch(
+            ActivityName => $Value,
+        );
+
+        if ( IsArrayRefWithData($ActivityEntityIDs) ) {
+
+            # Add search term from input field.
+            push @{$ActivityEntityIDs}, $Value;
+
+            # Return search parameter structure.
+            return {
+                $Operator => $ActivityEntityIDs,
+                }
+        }
+    }
+
+    return {
+        $Operator => $Value,
+    };
 }
 
 1;