+++ scripts/test/DynamicField/StatsSearchFieldParameterBuild.t
@@ -14,10 +14,82 @@ use vars (qw($Self));
 
 use Kernel::System::VariableCheck qw(:all);
 
-# get dynamic field backend object
 my $DFBackendObject = $Kernel::OM->Get('Kernel::System::DynamicField::Backend');
 
-my $UserID = 1;
+my $ActivityObject = $Kernel::OM->Get('Kernel::System::ProcessManagement::DB::Activity');
+my $ProcessObject  = $Kernel::OM->Get('Kernel::System::ProcessManagement::DB::Process');
+
+$Kernel::OM->ObjectParamAdd(
+    'Kernel::System::UnitTest::Helper' => {
+        RestoreDatabase => 1,
+    },
+);
+my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
+
+my $UserID   = 1;
+my $RandomID = $Helper->GetRandomID();
+
+my %ProcessLookup = (
+    'EntityID-1' . $RandomID => 'Process-1' . $RandomID,
+    'EntityID-2' . $RandomID => 'Process-2' . $RandomID,
+    'EntityID-3' . $RandomID => 'Process-3' . $RandomID,
+);
+my $ActivityEntityID       = 'A1-' . $RandomID;
+my $ActivityDialogEntityID = 'AD1-' . $RandomID;
+
+for my $Process ( sort keys %ProcessLookup ) {
+    my $ProcessID = $ProcessObject->ProcessAdd(
+        EntityID      => $Process,
+        Name          => $ProcessLookup{$Process},
+        StateEntityID => 'S1',
+        Layout        => {},
+        Config        => {
+            Description => 'a Description',
+            Path        => {
+                $ActivityEntityID => {},
+                }
+        },
+        UserID => $UserID,
+    );
+
+    $Self->True(
+        $ProcessID,
+        "Process is created - $ProcessID.",
+    );
+
+}
+
+my %ActivityLookup = (
+    'EntityID-1' . $RandomID => 'Activity-1' . $RandomID,
+    'EntityID-2' . $RandomID => 'Activity-2' . $RandomID,
+    'EntityID-3' . $RandomID => 'Activity-3' . $RandomID,
+);
+
+for my $Activity ( sort keys %ActivityLookup ) {
+    my $ActivityID = $ActivityObject->ActivityAdd(
+        EntityID => $Activity,
+        Name     => $ActivityLookup{$Activity},
+        Config   => {
+            ActivityDialog => {
+                1 => $ActivityDialogEntityID,
+            },
+        },
+        UserID => $UserID,
+    );
+
+    $Self->True(
+        $ActivityID,
+        "Activity is created - $ActivityID.",
+    );
+}
+
+my $DynamicFieldProcessID = $Kernel::OM->Get('Kernel::System::DynamicField')->DynamicFieldGet(
+    Name => 'ProcessManagementProcessID',
+);
+
+my $DynamicFieldActivityID = $Kernel::OM->Get('Kernel::System::DynamicField')->DynamicFieldGet(
+    Name => 'ProcessManagementActivityID',
+);
 
 # theres is not really needed to add the dynamic fields for this test, we can define a static
 # set of configurations
@@ -362,6 +434,106 @@ my @Tests = (
         },
         ExpectedResults => {
             GreaterThanEquals => '2015-05-01 02:15:00',
+        },
+    },
+    {
+        Name   => 'ProcessManagementProcessID DF tets 1',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldProcessID,
+            Value              => $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ sort keys %ProcessLookup, $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementProcessID DF tets 2',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldProcessID,
+            Value              => 'Process-1' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-1' . $RandomID, 'Process-1' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementProcessID DF tets 3',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldProcessID,
+            Value              => 'Process-2' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-2' . $RandomID, 'Process-2' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementProcessID DF tets 4',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldProcessID,
+            Value              => 'Process-3' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-3' . $RandomID, 'Process-3' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementProcessID DF tets 5',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldProcessID,
+            Value              => "*$RandomID",
+        },
+        ExpectedResults => {
+            Like => [ ( sort keys %ProcessLookup ), "*$RandomID" ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementActivityID DF tets 1',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldActivityID,
+            Value              => $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ sort keys %ActivityLookup, $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementActivityID DF tets 2',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldActivityID,
+            Value              => 'Activity-1' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-1' . $RandomID, 'Activity-1' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementActivityID DF tets 3',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldActivityID,
+            Value              => 'Activity-2' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-2' . $RandomID, 'Activity-2' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementActivityID DF tets 4',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldActivityID,
+            Value              => 'Activity-3' . $RandomID,
+        },
+        ExpectedResults => {
+            Equals => [ 'EntityID-3' . $RandomID, 'Activity-3' . $RandomID ],
+        },
+    },
+    {
+        Name   => 'ProcessManagementActivityID DF tets 5',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldActivityID,
+            Value              => "*$RandomID",
+        },
+        ExpectedResults => {
+            Like => [ ( sort keys %ActivityLookup ), "*$RandomID" ],
         },
     },
 );